Upgrade golangci-lint to v2.1.1

Former-commit-id: 5d5d5df31cb6f396902a4966533929ca79cf9fd3

+++ cmd/skywire-cli/commands/rewards/server.go
@@ -229,7 +229,7 @@ func server() {
 	case <-dmsgclient.Ready():
 	}
 
-	lis, err := dmsgclient.Listen(uint16(dmsgPort)) //nolint: gosec
+	lis, err := dmsgclient.Listen(dmsgPort) //nolint: gosec
 	if err != nil {
 		log.WithError(err).Fatal()
 	}
